# core/src/main/resources/swagger_spec.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0",
    "title" : "Klaw Core",
    "contact" : {
      "email" : "info@klaw-project.io"
    }
  },
  "basePath" : "/",
  "paths" : {
...
    "/deleteSchemaRequests" : {
      "post" : {
        "operationId" : "deleteSchemaRequests",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "req_no",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/execSchemaRequests" : {
      "post" : {
        "operationId" : "execSchemaRequests",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "avroSchemaReqId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/execSchemaRequestsDecline" : {
      "post" : {
        "operationId" : "execSchemaRequestsDecline",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "avroSchemaReqId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "reasonForDecline",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/getCreatedSchemaRequests" : {
      "get" : {
        "operationId" : "getCreatedSchemaRequests",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "currentPage",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "requestsType",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "default" : "created"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SchemaRequestModel"
              }
            }
          }
        }
      }
    },
...
    "/getSchemaRequests" : {
      "get" : {
        "operationId" : "getSchemaRequests",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "currentPage",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "requestsType",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "default" : "all"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SchemaRequestModel"
              }
            }
          }
        }
      }
    },
...
    "/uploadSchema" : {
      "post" : {
        "operationId" : "uploadSchema",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SchemaRequestModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "status" : {
          "type" : "string",
          "enum" : [ "CONTINUE", "SWITCHING_PROTOCOLS", "PROCESSING", "CHECKPOINT", "OK", "CREATED", "ACCEPTED", "NON_AUTHORITATIVE_INFORMATION", "NO_CONTENT", "RESET_CONTENT", "PARTIAL_CONTENT", "MULTI_STATUS", "ALREADY_REPORTED", "IM_USED", "MULTIPLE_CHOICES", "MOVED_PERMANENTLY", "FOUND", "MOVED_TEMPORARILY", "SEE_OTHER", "NOT_MODIFIED", "USE_PROXY", "TEMPORARY_REDIRECT", "PERMANENT_REDIRECT", "BAD_REQUEST", "UNAUTHORIZED", "PAYMENT_REQUIRED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "PROXY_AUTHENTICATION_REQUIRED", "REQUEST_TIMEOUT", "CONFLICT", "GONE", "LENGTH_REQUIRED", "PRECONDITION_FAILED", "PAYLOAD_TOO_LARGE", "REQUEST_ENTITY_TOO_LARGE", "URI_TOO_LONG", "REQUEST_URI_TOO_LONG", "UNSUPPORTED_MEDIA_TYPE", "REQUESTED_RANGE_NOT_SATISFIABLE", "EXPECTATION_FAILED", "I_AM_A_TEAPOT", "INSUFFICIENT_SPACE_ON_RESOURCE", "METHOD_FAILURE", "DESTINATION_LOCKED", "UNPROCESSABLE_ENTITY", "LOCKED", "FAILED_DEPENDENCY", "UPGRADE_REQUIRED", "PRECONDITION_REQUIRED", "TOO_MANY_REQUESTS", "REQUEST_HEADER_FIELDS_TOO_LARGE", "UNAVAILABLE_FOR_LEGAL_REASONS", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "GATEWAY_TIMEOUT", "HTTP_VERSION_NOT_SUPPORTED", "VARIANT_ALSO_NEGOTIATES", "INSUFFICIENT_STORAGE", "LOOP_DETECTED", "BANDWIDTH_LIMIT_EXCEEDED", "NOT_EXTENDED", "NETWORK_AUTHENTICATION_REQUIRED" ]
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time"
        },
        "message" : {
          "type" : "string"
        },
        "debugMessage" : {
          "type" : "string"
        },
        "result" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        }
      }
    },
...
    "SchemaRequestModel" : {
      "type" : "object",
      "properties" : {
        "req_no" : {
          "type" : "integer",
          "format" : "int32"
        },
        "topicname" : {
          "type" : "string"
        },
        "environment" : {
          "type" : "string"
        },
        "environmentName" : {
          "type" : "string"
        },
        "schemaversion" : {
          "type" : "string"
        },
        "teamname" : {
          "type" : "string"
        },
        "teamId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "appname" : {
          "type" : "string"
        },
        "schemafull" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "requesttime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "requesttimestring" : {
          "type" : "string"
        },
        "topicstatus" : {
          "type" : "string"
        },
        "requesttype" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        },
        "approver" : {
          "type" : "string"
        },
        "approvingtime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "approvingTeamDetails" : {
          "type" : "string"
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "currentPage" : {
          "type" : "string"
        }
      }
    },
...
  }
}